% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{Gelman}
\alias{Gelman}
\title{Calculate the Gelman & Ruben Diagnostic Ratio}
\usage{
Gelman(x)
}
\arguments{
\item{x}{Either a matrix or a mcmc.list}
}
\description{
Calculate the Gelman & Ruben Diagnostic Ratio
}
\examples{
# Target Distribution
df <- function(x){
   return(.7*dnorm(x, mean=2, sd=1) + .3*dnorm(x, mean=5, sd=1))
}

# Proposal distribution
rprop <- function(x){ x + runif(1, -2, 2)}
dprop <- function(x.star, x){ dunif(x.star-x, -2, 2) }

# Create chains
chain1 <- MCMC(df, start= 4, rprop, dprop, N=1000)
chain2 <- MCMC(df, start= 8, rprop, dprop, N=1000)
chain3 <- MCMC(df, start=-3, rprop, dprop, N=1000)
chain4 <- MCMC(df, start=-6, rprop, dprop, N=1000)
chains <- cbind(chain1, chain2, chain3, chain4)
Gelman(chains)

# Or using mMCMC
chains <- mMCMC(df, start=list(4,8,-3,-6), rprop, dprop, N=1000, num.chains=4)
traceplot(chains)
chains <- window(chains, 100, 1000)
Gelman(chains)
}

