% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{Create a single Markov Chain Monte Carlo simulation}
\usage{
MCMC(df, start, rprop, dprop = function(to, from) {     1 }, N = 1000)
}
\arguments{
\item{df}{A function that is the target distribution}

\item{start}{The starting point for the MCMC}

\item{rprop}{A function that takes the current chain step and proposes a new value}

\item{dprop}{The density function of the proposal distribution. If NULL, it is assumed to be symetric.}

\item{N}{How many chain steps to calculate.}
}
\description{
Create a single Markov Chain Monte Carlo simulation
}
\examples{
# Target Distribution
df <- function(x){
   return(.7*dnorm(x, mean=2, sd=1) + .3*dnorm(x, mean=5, sd=1))
}

# Proposal distribution
rprop <- function(x){ x + runif(1, -2, 2)}
dprop <- function(x.star, x){ dunif(x.star-x, -2, 2) }

# Create 1 chain
chain1 <- MCMC(df, start=4, rprop, dprop, N=1000)
trace_plot(chain1)

# Create several chains
chain2 <- MCMC(df, start= 8, rprop, dprop, N=1000)
chain3 <- MCMC(df, start=-3, rprop, dprop, N=1000)
chain4 <- MCMC(df, start=-6, rprop, dprop, N=1000)
chains <- cbind(chain1, chain2, chain3, chain4)
trace_plot(chains)
}

