% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{plot_2D_chains}
\alias{plot_2D_chains}
\title{Plot MCMC chains along a 2D density}
\usage{
plot_2D_chains(target.map, chains)
}
\arguments{
\item{target.map}{A density plot of a 2-D distribution}

\item{chains}{An mcmc.list}
}
\description{
Plot MCMC chains along a 2D density
}

