% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{plot_chain}
\alias{plot_chain}
\title{Graph the first few steps of a MCMC chain}
\usage{
plot_chain(density, chain, proposed)
}
\arguments{
\item{density}{A ggplot2 object that shows a density function}

\item{chain}{A vector of values of the chain}

\item{proposed}{A vector of proposed values for the chain}
}
\description{
Graph the first few steps of a MCMC chain
}

