% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{plot_HMC_proposal_path}
\alias{plot_HMC_proposal_path}
\title{Plot a Hamiltonian proposal path}
\usage{
plot_HMC_proposal_path(target.map, steps)
}
\arguments{
\item{target.map}{A density plot of a 2-D distribution}

\item{steps}{The output of HMC.one.step}
}
\description{
Plot a Hamiltonian proposal path
}

