% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MCMC.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a traceplot from a MCMC chain or chains}
\usage{
trace_plot(chain)
}
\arguments{
\item{chain}{Either a mcmc.list or a matrix or a vector}
}
\description{
Create a traceplot from a MCMC chain or chains
}

